refactor: data 삽입 검증 로직 삭제
