+++ src/main/java/com/ray/pominowner/global/config/category/CategoryCache.java
@@ -1,22 +1,36 @@
 package com.ray.pominowner.global.config.category;
 
 import com.ray.pominowner.store.domain.Category;
+import com.ray.pominowner.store.repository.CategoryRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Component;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
-public final class CategoryCache {
+@Slf4j
+@RequiredArgsConstructor
+@Component
+@Profile("dev")
+public final class CategoryCache implements ApplicationRunner {
 
-    private static final List<Category> INITIAL_CATEGORY_CACHE;
+    private final CategoryRepository repository;
 
-    static {
-        INITIAL_CATEGORY_CACHE = Arrays.stream(InitialCategoryInfo.values())
-                .map(initialCategoryInfo -> new Category(initialCategoryInfo.name(), initialCategoryInfo.url()))
-                .collect(Collectors.toList());
+    private static final List<Category> INITIAL_CATEGORY_CACHE = new ArrayList<>();
+
+    @Override
+    public void run(final ApplicationArguments args) throws Exception {
+        log.info("-----------caching initial category list-----------");
+        List<Category> initialCategories = repository.findAll();
+
+        INITIAL_CATEGORY_CACHE.addAll(initialCategories);
     }
 
-    public static List<Category> cachedCategoryList() {
+    public static List<Category> initialCategories() {
         return List.copyOf(INITIAL_CATEGORY_CACHE);
     }
